pyproject.toml
[project]
name = "medinodus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crypto>=1.4.1",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "psycopg2",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "supabase>=2.28.3",
    "uvicorn>=0.43.0",
]
